\name{blkdiag}

\alias{blkdiag}
%- Also NEED an '\alias' for EACH other topic documented here.

\title{Build a block diagonal matrix.}

\description{Function to build a block diagonal matrix (internal function).}

\usage{blkdiag(A, B)}

%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{A}{The first non-empty, numeric matrix}
  \item{B}{The second non-empty, numeric matrix}
}

\value{
  \item{C}{A numeric block diagonal matrix}

}

\author{Francesco Bartolucci, Silvia Pandolfi - University of Perugia (IT)}
