\name{data_SRHS}

\alias{data_SRHS}

\docType{data}

\title{Self-reported health status dataset}

\description{Dataset about self-reported health status derived from the Health and Retirement Study conducted by the University of Michigan.}

\usage{data(data_SRHS)}

\format{
  A data frame with 7074 observations on the following 19 variables.
  \describe{
    \item{\code{R1SHLT}}{self-reported health status at the 1st time occasion}
    \item{\code{R2SHLT}}{self-reported health status at the 2nd time occasion}
    \item{\code{R3SHLT}}{self-reported health status at the 3rd time occasion}
    \item{\code{R4SHLT}}{self-reported health status at the 4th time occasion}
    \item{\code{R5SHLT}}{self-reported health status at the 5th time occasion}
    \item{\code{R6SHLT}}{self-reported health status at the 6th time occasion}
    \item{\code{R7SHLT}}{self-reported health status at the 7th time occasion}
    \item{\code{R8SHLT}}{self-reported health status at the 8th time occasion}
    \item{\code{RAGENDER}}{gender}
    \item{\code{RARACEM}}{race}
    \item{\code{RAEDUC}}{educational level}
    \item{\code{R1AGEY_E}}{age at the 1st time occasion}
    \item{\code{R2AGEY_E}}{age at the 2nd time occasion}
    \item{\code{R3AGEY_E}}{age at the 3rd time occasion}
    \item{\code{R4AGEY_E}}{age at the 4th time occasion}
    \item{\code{R5AGEY_E}}{age at the 5th time occasion}
    \item{\code{R6AGEY_E}}{age at the 6th time occasion}
    \item{\code{R7AGEY_E}}{age at the 7th time occasion}
    \item{\code{R8AGEY_E}}{age at the 8th time occasion}
  }
}

\references{Bartolucci, F., Bacci, S. and Pennoni, F. (2013), Longitudinal analysis of the self-reported health status by mixture latent autoregressive models, Journal of the Royal Statistical Society - series C, in press (previous technical report: http://arxiv.org/abs/1108.1498)}

\examples{data(data_SRHS)}

\keyword{datasets}
