\name{expit}

\alias{expit}
%- Also NEED an '\alias' for EACH other topic documented here.

\title{Compute the expit function.}

\description{Function to compute the inverse of the logit function.}

\usage{expit(x)}

%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{a vector of matrix which contains the values to be transformed}
}

\value{
  \item{y}{The transformed sample}

}

\author{Francesco Bartolucci, Silvia Pandolfi - University of Perugia (IT)}
