\name{invglob}

\alias{invglob}
%- Also NEED an '\alias' for EACH other topic documented here.

\title{Invert vector of global logits.}

\description{Function to invert the marginal parametrization based on global logits (internal funtion).}

\usage{invglob(eta)}

%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{eta}{vector of global logits}
}

\value{
  \item{p}{vector of joint probabiltie}
}

\author{Francesco Bartolucci, University of Perugia, http://www.stat.unipg.it/bartolucci}
