\name{logit1}

\alias{logit1}
%- Also NEED an '\alias' for EACH other topic documented here.

\title{Compute the logit function with respect to a reference category.}

\description{Function to compute the logit function with respect to a reference category.}

\usage{logit1(p, ref = 1)}

%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{p}{a vector which contains the values to be transformed}
  \item{ref}{the reference category}
}

\value{
  \item{lp}{The transformed vector}

}

\author{Francesco Bartolucci, Silvia Pandolfi - University of Perugia (IT)}
