\name{prod_array}

\alias{prod_array}
%- Also NEED an '\alias' for EACH other topic documented here.

\title{Compute the product of array and vector}

\description{Function that compute the product of each matrix in X by the vector y  (internal funtion).}

\usage{prod_array(X, y)}

%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{X}{a given array}
  \item{y}{a given vector}
}

\value{
\item{Z}{the resulting matrix}
}

\author{Francesco Bartolucci, Silvia Pandolfi, University of Perugia (IT), http://www.stat.unipg.it/bartolucci}
