\name{sq}
\alias{sq}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Create a matrix with the combination of vectors of (1,0) }
\description{Function that creates a matrix with the combination of all possible row vectors with 't' elements equal to 1 and 'J-t' elements equal to 0. If 't' is NULL the function generates all the possible combinations of vectors of length 'J' with elements (1,0).}
\usage{sq(J,t = NULL)}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{J}{an integer} 
  \item{t}{an integer}
 
 
}
\value{
 \item{M}{the resulting matrix}
 }
 
\author{Francesco Bartolucci, Silvia Pandolfi - University of Perugia (IT)}
