\name{data_SRHS_long}

\alias{data_SRHS_long}

\docType{data}

\title{Self-reported health status dataset}

\description{Dataset about self-reported health status derived from the Health and Retirement Study conducted by the University of Michigan.}

\usage{data(data_SRHS_long)}

\format{
  A data frame with 56592 observations on the following 6 variables.
  \describe{
    \item{\code{id}}{subject id}
    \item{\code{gender}}{gender of the subject}
    \item{\code{race}}{race}
    \item{\code{education}}{educational level}
    \item{\code{age}}{age at the different time occasions}
    \item{\code{srhs}}{self-reported health status at the different time occasions}
  }
}

\references{Bartolucci, F., Bacci, S. and Pennoni, F. (2014), Longitudinal analysis of the self-reported health status by mixture latent autoregressive models, Journal of the Royal Statistical Society - series C, 63, pp. 267-288}

\examples{data(data_SRHS_long)}

\keyword{datasets}
