\name{summary}
\alias{summary.LMbasic}
\alias{summary.LMbasiccont}
\alias{summary.LMlatent}
\alias{summary.LMlatentcont}
\alias{summary.LMmanifest}
\alias{summary.LMmixed}
\alias{summary.MCbasic}
\alias{summary.MCcov}
\alias{summary.LMsearch}

\title{Summary of LM fits}
\description{Summary methods}

\usage{
\method{summary}{LMbasic}(object, ...)
\method{summary}{LMbasiccont}(object, ...)
\method{summary}{LMlatent}(object, ...)
\method{summary}{LMlatentcont}(object, ...)
\method{summary}{LMmanifest}(object, ...)
\method{summary}{LMmixed}(object, ...)
\method{summary}{MCbasic}(object, ...)
\method{summary}{MCcov}(object, ...)
\method{summary}{LMsearch}(object,...)

}


\arguments{
\item{object}{output from \code{\link{lmest}},\code{\link{lmestCont}},\code{\link{lmestMixed}}, and \code{\link{lmestMc}}}
\item{...}{further arguments passed to or from other methods}
}

\author{Francesco Bartolucci,
  Silvia Pandolfi,
  Fulvia Pennoni,
  Alessio Farcomeni,
  Alessio Serafini
}
